chore: move node definition classes
