+++ src/shaderityShaders/Input/Geometry/inputPosition.vs.glsl
