+++ src/shaderityShaders/ShaderPack.ts
@@ -1,21 +1,21 @@
-import uniformTexture2D from './common/uniformTexture2D';
-
 import {AbstractNodeData} from './../type/Type';
-import uniformVector4 from './common/uniformVector4.glsl';
-import uniformColor from './common/uniformColor.glsl';
-import uniformMatrix44 from './common/uniformMatrix44.glsl';
-import sampleTexture2D from './common/sampleTexture2D.glsl';
 
-import inputPosition from './vertex/inputPosition.vs.glsl';
-import vertexPosition from './vertex/vertexPosition.vs.glsl';
-import texcoord from './vertex/texcoord.glsl';
+import uniformVector4 from './Input/Basic/uniformVector4.glsl';
 
-import fragmentColor from './fragment/fragmentColor.fs.glsl';
+import inputPosition from './Input/Geometry/inputPosition.vs.glsl';
+import texcoord from './Input/Geometry/texcoord.glsl';
+
+import uniformMatrix44 from './Input/Matrix/uniformMatrix44.glsl';
+
+import sampleTexture2D from './Input/Texture/sampleTexture2D.glsl';
+import uniformTexture2D from './Input/Texture/uniformTexture2D';
+
+import fragmentColor from './Utility/Output/fragmentColor.fs.glsl';
+import vertexPosition from './Utility/Output/vertexPosition.vs.glsl';
 
 export const ShaderPack: AbstractNodeData[] = [
   uniformTexture2D,
   uniformVector4,
-  uniformColor,
   uniformMatrix44,
   sampleTexture2D,
   inputPosition,
