+++ src/shaderityShaders/Utility/Output/fragmentColor.fs.glsl
